Finish renaming, add a exc_wrap_vsilfile()
