+++ rasterio/_err.pyx
@@ -193,3 +193,16 @@ cdef void *exc_wrap_pointer(void *ptr) except NULL:
             raise exc
             return NULL
     return ptr
+
+
+cdef VSILFILE *exc_wrap_vsilfile(VSILFILE *f) except NULL:
+    """Wrap a GDAL/OGR function that returns GDALDatasetH etc (void *)
+
+    Raises a Rasterio exception if a non-fatal error has be set.
+    """
+    if f == NULL:
+        exc = exc_check()
+        if exc:
+            raise exc
+            return NULL
+    return f
